backend/requirements.txt
transformers==4.35.0
torch>=2.6.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
Pillow==10.0.1
celery==5.3.0
redis==5.0.1